use new indexer files structureclear
